added logic for storage enitity and clean up code

+++ src/main/java/ua/grainmole/dto/GroundSectionDto.java
@@ -2,6 +2,8 @@ package ua.grainmole.dto;
 
 import lombok.Builder;
 
+import java.math.BigInteger;
+
 @Builder
-public record GroundSectionDto(Integer groundSectionId, Integer userId) {
+public record GroundSectionDto(BigInteger groundSectionId, Integer userId) {
 }
